Increase delta precision

+++ src/engine/Time.java
@@ -8,16 +8,16 @@ public class Time {
 
     private static long lastLoopTime = nanoTime();
 
-    private static float delta;
+    private static double delta;
 
 
     public static void calculateDelta() {
         long time = nanoTime();
-        delta = (float) (time - lastLoopTime) / 1000000f;
+        delta = (time - lastLoopTime) / 1000000d;
         lastLoopTime = time;
     }
 
-    public static float getDelta() {
+    public static double getDelta() {
         return(delta);
     }
 }
